rewrite all the GDScript in C#
if, for some reason, you are reading this commit message: do not do this yourself. this was an absurd thing to do because GDScript was making me sad and I wanted to refresh my C# knowledge. this isn't even a good rewrite, it's largely a direct translation of the old code, warts and all. probably going to re-rewrite most of this anyway. c'est la vie!